Added Test Bed mocks for Firebase, moved services to separate folders
