+++ src/app/home/problem-list/problem-list.component.html
@@ -1,32 +1,12 @@
-<html>
-  <head>
-    <style>
-      table,
-      th,
-      td {
-        border: 1px solid black;
-      }
-
-      th,
-      td {
-        padding: 5px;
-        text-align: left;
-      }
-    </style>
-  </head>
-
-  <body>
-    <table>
-      <thead>
-        <th>Title</th>
-        <th>Description</th>
-      </thead>
-      <tbody>
-        <tr *ngFor="let problem of problems">
-          <td>{{ problem.title }}</td>
-          <td>{{ problem.description }}</td>
-        </tr>
-      </tbody>
-    </table>
-  </body>
-</html>
+<table>
+  <thead>
+    <th>Title</th>
+    <th>Description</th>
+  </thead>
+  <tbody>
+    <tr *ngFor="let problem of problems">
+      <td>{{ problem.title }}</td>
+      <td>{{ problem.description }}</td>
+    </tr>
+  </tbody>
+</table>
