+++ src/app/home/problem-list/problem-list.component.scss
@@ -1,0 +1,11 @@
+table,
+th,
+td {
+    border: 1px solid black;
+}
+
+th,
+td {
+    padding: 5px;
+    text-align: left;
+}
