+++ src/app/home/problem-list/problem-list.component.spec.ts
@@ -17,8 +17,4 @@ describe('ProblemListComponent', () => {
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
 });
