+++ src/app/home/problem-list/problem.model.ts
@@ -1,4 +1,4 @@
-export interface Problem {
+export interface IProblem {
   id: string;
   title: string;
   description: string;
