+++ src/app/services/utility-services/problem-list.service.ts
@@ -1,6 +1,5 @@
 import { Injectable } from '@angular/core';
 import { AngularFirestore } from '@angular/fire/firestore';
-import { Problem } from 'src/app/home/problem-list/problem.model';
 
 @Injectable({
   providedIn: 'root'
